backend/requirements.txt
langchain>=1.2.0
langchain-classic>=1.0.0
langchain-core>=1.2.0
langchain-community>=0.3.28
langchain-anthropic>=1.4.0
langchain-openai>=1.1.0
langchain-chroma>=1.0.0
langchain-text-splitters>=1.1.0
chromadb>=1.5.0
pypdf>=4.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
